+++ CHANGES.md
@@ -12,6 +12,8 @@
 ## master
 
 - [CHANGE] ディレクトリ構成を変更して、`<ggrpc/ggrpc.h>` をインクルードして使うようにした
+- [CHANGE] クライアントの `OnReadDone`, `OnResponse` を `OnFinish` に変更
+- [CHANGE] `ClientResponseReader` の `Request` を `Connect` に変更
 - [ADD] クライアントストリーミングを実装
 
 ## 0.3.0